refactor: create insomnia-data and move database files in
